bottom navbar functionality done
