+++ lib/assesments/assessment_2/constants/f1_as2_amna_colors.dart
@@ -1,0 +1,5 @@
+import 'package:flutter/material.dart';
+
+class F1As2AmnaConstants {
+  static const primaryColor = Color(0xFFFA9884);
+}
