+++ lib/assesments/assessment_2/screens/f1_as2_amna_notification.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+class F1As2AmnaNotification extends StatelessWidget {
+  const F1As2AmnaNotification({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Scaffold(
+      body: Center(child: Text("Notification")),
+    );
+  }
+}
