+++ lib/assesments/assessment_2/screens/f1_as2_amna_search.dart
@@ -1,0 +1,14 @@
+import 'package:flutter/material.dart';
+
+class F1As2AmnaSearch extends StatelessWidget {
+  const F1As2AmnaSearch({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Scaffold(
+      body: Center(
+        child: Text("Search"),
+      ),
+    );
+  }
+}
