+++ lib/assesments/assessment_2/screens/profile/f1_as2_amna_profile.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+class F1As2AmnaProfile extends StatelessWidget {
+  const F1As2AmnaProfile({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Scaffold(
+      body: Center(child: Text("Profile")),
+    );
+  }
+}
