+++ pubspec.yaml
@@ -60,7 +60,6 @@ flutter:
   # To add assets to your application, add an assets section, like this:
   assets:
     - assets/
-    - assets/amna/svgs/
     - assets/svgs/
     - assets/amna/
